Separate search API into its own file
