+++ src/api/api.js
@@ -5,6 +5,7 @@ import * as LocationClientIsp from './calls/locationClientIsp';
 import * as LocationClientIspTransitIsp from './calls/locationClientIspTransitIsp';
 import * as LocationTransitIsp from './calls/locationTransitIsp';
 import * as TransitIsp from './calls/transitIsp';
+import * as Search from './calls/search';
 
 console.warn('Multipart API queries (e.g. locationTransitIsp) not using proper URLs and have incorrect transforms.');
 
@@ -16,6 +17,7 @@ const calls = {
   ...LocationClientIspTransitIsp,
   ...LocationTransitIsp,
   ...TransitIsp,
+  ...Search,
 };
 
 export default calls;
