+++ src/api/calls/clientIsp.js
@@ -11,17 +11,6 @@ import {
 } from '../transforms';
 
 /**
- * Get Search results for a client ISP
- *
- * @param {String} searchQuery search to search for. (e.g. comcas)
- * @return {Promise} A promise after the get request was made
- */
-export function getClientIspSearch(searchQuery) {
-  return get('/clients/search', { q: stringToKey(searchQuery) })
-    .then(transform(transformClientIspSearchResults));
-}
-
-/**
  * Get information for a client ISP
  *
  * @param {String} clientIspId The client ISP to query (e.g., AS7922)
